requirements.txt
asgiref==3.3.1
boto3==1.16.37
botocore==1.19.37
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.43.0
django-countries==7.0
django-crispy-forms==1.10.0
django-storages==1.10.1
gunicorn==20.0.4
jmespath==0.10.0
oauthlib==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.4
requests-oauthlib==1.3.0
s3transfer==0.3.3
sqlparse==0.4.1
stripe==2.55.1
